MTDSA-23692:  API#1885 - Update fields to contain only whole numbers (#512)
* MTDSA-23692: updated code and ITs

* MTDSA-23692: Updated OAS

* [MTDSA-23693] [CK] Use BigInt instead of Long to over come the SBT compilation bug.

---------

+++ app/v5/retrieveCalculation/def2/model/response/calculation/transitionProfit/TransitionProfit.scala
@@ -19,7 +19,7 @@ package v5.retrieveCalculation.def2.model.response.calculation.transitionProfit
 import play.api.libs.json.{Format, Json}
 
 case class TransitionProfit(
-    totalTaxableTransitionProfit: Option[BigDecimal],
+    totalTaxableTransitionProfit: Option[BigInt],
     transitionProfitDetail: Option[Seq[TransitionProfitDetail]]
 )
